Move serializer
